Fix pre test CI
